+++ packages/cli/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Patch Changes
 
-- e4313d4: Initial publish to npm under @flusk org.
-- Updated dependencies [e4313d4]
+- [`e4313d4`](https://github.com/adirbenyossef/flusk-dev/commit/e4313d4dea191d253d5d92ad29438ccbc3690629) Thanks [@adirbenyossef](https://github.com/adirbenyossef)! - Initial publish to npm under @flusk org.
+
+- Updated dependencies [[`e4313d4`](https://github.com/adirbenyossef/flusk-dev/commit/e4313d4dea191d253d5d92ad29438ccbc3690629)]:
   - @flusk/forge@0.1.1
   - @flusk/logger@0.1.1
   - @flusk/entities@0.1.1