+++ packages/logger/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Patch Changes
 
-- e4313d4: Initial publish to npm under @flusk org.
+- [`e4313d4`](https://github.com/adirbenyossef/flusk-dev/commit/e4313d4dea191d253d5d92ad29438ccbc3690629) Thanks [@adirbenyossef](https://github.com/adirbenyossef)! - Initial publish to npm under @flusk org.